requirements.txt
torch>=2.0.0
numpy>=1.24.0
wandb>=0.15.0
hydra-core>=1.3.0
scikit-learn>=1.3.0
